cleaned up test file by creating config file
