+++ testudo-module/test/test.js
@@ -1,40 +1,13 @@
 import fs from 'fs';
 import lighthouse from 'lighthouse';
-import * as chromeLauncher from 'chrome-launcher';
 // imports the desktop-specific parameters for config
-import { throttling, screenEmulation, emulatedUserAgent } from './constants.js';
+import config from './config.js';
+const { desktopConfig, options, chrome } = config;
 
-const chrome = await chromeLauncher.launch({
-  chromeFlags: ['--headless'],
-});
-const options = {
-  logLevel: 'info',
-  output: 'html', // <-- change output to html/json as needed
-  onlyCategories: ['performance'],
-  port: chrome.port,
-};
-// used as 3rd parameter for lighthouse function, desktop
-const config = {
-  extends: 'lighthouse:default',
-  settings: {
-    maxWaitForFcp: 15 * 1000,
-    maxWaitForLoad: 35 * 1000,
-    formFactor: 'desktop',
-    throttling,
-    screenEmulation,
-    emulatedUserAgent,
-    skipAudits: [
-      // Skip the h2 audit so it doesn't lie to us. See https://github.com/GoogleChrome/lighthouse/issues/6539
-      'uses-http2',
-      // There are always bf-cache failures when testing in headless. Reenable when headless can give us realistic bf-cache insights.
-      'bf-cache',
-    ],
-  },
-};
 const runnerResult = await lighthouse(
   'https://www.codesmith.io',
   options,
-  config
+  desktopConfig
 );
 
 // `.report` is the HTML report as a string
